requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
azure-core==1.30.2
azure-identity==1.17.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
Deprecated==1.2.14
exceptiongroup==1.2.2
frozenlist==1.4.1
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.4.0
microsoft-kiota-abstractions==1.3.3
microsoft-kiota-authentication-azure==1.1.0
microsoft-kiota-http==1.3.3
microsoft-kiota-serialization-form==0.1.1
microsoft-kiota-serialization-json==1.3.1
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
msal==1.30.0
msal-extensions==1.2.0
msgraph-core==1.1.3
msgraph-sdk==1.6.0
multidict==6.0.5
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
outcome==1.3.0.post0
packaging==24.1
pendulum==3.0.0
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
std-uritemplate==1.0.6
time-machine==2.15.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
websocket-client==1.8.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.11
zipp==3.20.1
